modify multi_mi_sim by using pair_sim instance
