+++ sdgx/metrics/multi_table/multitable_mi_sim.py
@@ -7,16 +7,10 @@ from scipy.stats import entropy
 from sklearn.metrics.cluster import normalized_mutual_info_score
 
 from sdgx.metrics.multi_table.base import MultiTableMetric
+from sdgx.metrics.pair_column.mi_sim import MISim
 
 
-def Jaccard_index(A, B):
-    return min(A, B) / max(A, B)
 
-
-def time2int(datetime, form):
-    time_array = time.strptime(datetime, form)
-    time_stamp = int(time.mktime(time_array))
-    return time_stamp
 
 
 class MISim(MultiTableMetric):
@@ -44,8 +38,8 @@ class MISim(MultiTableMetric):
         Calculate the JSD value between a real column and a synthetic column.
         Args:
             real_data (pd.DataFrame): The real data.
-
             synthetic_data (pd.DataFrame): The synthetic data.
+            metadata(dict): The metadata that describes the data type of each column
 
         Returns:
             MI_similarity (float): The metric value.
@@ -55,36 +49,17 @@ class MISim(MultiTableMetric):
 
         columns = synthetic_data.columns
         n = len(columns)
-
-        for col in columns:
-            data_type = metadata[col]
-            if data_type == "numerical":
-                # max_value = real_data[col].max()
-                # min_value = real_data[col].min()
-                real_data[col] = pd.cut(a, self.numerical_bins, labels=range(self.numerical_bins))
-                synthetic_data[col] = pd.cut(
-                    a, self.numerical_bins, labels=range(self.numerical_bins)
-                )
-
-            elif data_type == "datetime":
-                real_data[col] = real_data[col].apply(time2int)
-                synthetic_data[col] = synthetic_data[col].apply(time2int)
-                real_data[col] = pd.cut(a, self.numerical_bins, labels=range(self.numerical_bins))
-                synthetic_data[col] = pd.cut(
-                    a, self.numerical_bins, labels=range(self.numerical_bins)
-                )
-
+        mi_sim_instance = MISim()
         nMI_sim = np.zeros((n, n))
 
         for i in range(len(columns)):
             for j in range(len(columns)):
-                syn_MI_ij = normalized_mutual_info_score(
-                    synthetic_data[columns[i]], synthetic_data[columns[j]]
+                syn_data = pd.concat([synthetic_data[columns[i]], synthetic_data[columns[j]]], axis=1)
+                real_data = pd.concat([real_data[columns[i]], real_data[columns[j]]], axis=1)
+
+                nMI_sim[i][j] = mi_sim_instance.calculate(
+                    real_data, syn_data ,metadata
                 )
-                real_MI_ij = normalized_mutual_info_score(
-                    real_data[columns[i]], real_data[columns[j]]
-                )
-                nMI_sim[i][j] = Jaccard_index(syn_MI_ij, real_MI_ij)
 
         MI_sim = np.sum(nMI_sim) / n / n
         # test
